Modified Id setter database

+++ app/src/main/java/ium/unisa/stopsmoking/db/Goal.java
@@ -2,21 +2,21 @@ package ium.unisa.stopsmoking.db;
 
 public class Goal {
 
-    private int id;
+    private long id;
     private String name;
     private double price;
 
-    public Goal(int id, String name, double price) {
+    public Goal(long id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
